A new way to create a link, possibly creating a zettel too
When writing I often want to create a link to a new zettel without
switching to it immediately.  This new function will let you create a
link to an existing zettel, or to a new zettel after creating it.

Essentially this is a replacement for/combination of
neuron-insert-zettel-link and neuron-insert-new-zettel.