Replace an undefined function with a cognate
The `map-put!` function is not defined in Emacs 26.3 with recent
copies of `f`, `s`, `dash`, and `markdown-mode` installed. However,
`map-put` is, and seems to maintain the intended effects. Namely, with
this modification calling the function `neuron-toggle-connection-type`
on a link toggles whether `?cf` is included at the end of the link.